% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{cdOptim}
\alias{cdOptim}
\title{Coordinate descent optimisation for non smooth functions}
\usage{
cdOptim(theta, fn, theta_it, tol = 1e-06, maxit = 1000)
}
\arguments{
\item{theta}{Initial value.}

\item{fn}{Function to optimise.}

\item{theta_it}{Iterative theta for coordinate descent algorithm. Default
is NULL.}

\item{tol}{Relative convergence tolerance}

\item{maxit}{Maximum number of iterations if relative convergence is not
reached. Default is 1000.}
}
\value{
The function returns a list with following components
\item{par}{The best set of parameters}
\item{value}{The value of the function corresponding to best set of parameters.}
\item{iteration}{Number of iterations taken}
}
\description{
Function for minimising piecewise differentiable objective functions.
}
