% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{nonsmoothOptim}
\alias{nonsmoothOptim}
\title{Optimisation tool for non smooth functions}
\usage{
nonsmoothOptim(theta, fn, df, method = "SG", tol = 1e-05,
  maxit = 1000, h = 0.001, prox = NULL, acl = FALSE,
  theta_it = NULL)
}
\arguments{
\item{theta}{Initial value.}

\item{fn}{Function to optimise.}

\item{df}{Subgradient of the objective function.}

\item{method}{Method to be used for optimisation. Default is subgradient
method.}

\item{tol}{Relative convergence tolerance}

\item{maxit}{Maximum number of iterations if relative convergence is not
reached. Default is 1000.}

\item{h}{Stepsize for updating the set of parameters. Default is 0.001.}

\item{prox}{Proximal operator for proximal gradient method. Default is
NULL.}

\item{acl}{Flag for accelerated proximal gradient method. Default is false.}

\item{theta_it}{Iterative theta for coordinate descent algorithm. Default
is NULL.}
}
\value{
The function returns a list with following components
\item{par}{The best set of parameters}
\item{value}{The value of the function corresponding to best set of parameters.}
\item{iteration}{Number of iterations taken}
\item{method}{The method used for the optimisation}
}
\description{
Function for minimising piecewise differentiable objective functions.
}
