% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{pgOptim}
\alias{pgOptim}
\title{Proximal gradient optimisation for non smooth functions}
\usage{
pgOptim(theta, fn, df, prox, acl = F, tol = 1e-06, t)
}
\arguments{
\item{theta}{Initial value.}

\item{fn}{Function to optimise.}

\item{df}{Subgradient of the objective function.}

\item{prox}{Proximal operator for proximal gradient method.}

\item{acl}{Flag for accelerated proximal gradient method. Default is false.}

\item{tol}{Relative convergence tolerance}

\item{t}{Stepsizes for updating the set of parameters.}
}
\value{
The function returns a list with following components
\item{par}{The best set of parameters}
\item{value}{The value of the function corresponding to best set of parameters.}
\item{iteration}{Number of iterations taken}
}
\description{
Function for minimising piecewise differentiable objective functions.
}
