% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{stepsize}
\alias{stepsize}
\title{Sequence of step sizes for optimization.}
\usage{
stepsize(h = 0.01, m = 1000, n = 1000)
}
\arguments{
\item{h}{Starting value. Default is 0.01}

\item{m}{Number of constant steps. Default is 1000}

\item{n}{Number of diminishing steps. Default is 1000}
}
\value{
The sequence of stepsize
}
\description{
Function to generate stepsize for sub-gradient optimization and proximal-gradient optimization
}
