% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{nonsmoothOptim}
\alias{nonsmoothOptim}
\alias{nonsmoothOptim-package}
\title{nonsmoothOptim: This package is intended for optimisation
with piecewise differentiable function.}
\description{
This is an optimisation package for piecewise differentiable
functions. The package comprise of three different optimisation
techiniques: Subgradient, Proximal gradient and Coordinate
descent. There is also a helper function to generate stepsizes
for subgradient and proximal gradient optimisation.
}
\section{nonsmoothOptim functions}{

Use ls("package:nonsmoothOptim") to check the available functions.
}

